+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.36] - Unreleased
+
+### :toolbox: Added
+
+- Enhance Add-DiaNodeIcon cmdlet with new parameters:
+  - TableBorderColor (Hex color)
+  - TableBorder (1-10)
+  - TableStyle (solid, dashed, dotted, bold)
+
+### :arrows_clockwise: Changed
+
+- Increment module version to `0.2.36`
+
 ## [0.2.35] - 2025-11-16
 
 ### :toolbox: Added